agent/pyproject.toml
[project]
name = "cs323-interview-agent"
version = "0.1.0"
description = "LiveKit voice agent with Tavus avatar for CS 323 reading interviews"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[anthropic,deepgram,cartesia,silero,tavus]~=1.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "livekit-plugins-elevenlabs~=1.4",
]
